extract/requirements.txt
python-dotenv
requests
psycopg2-binary
boto3